Add basic scoring logic and display of it
The basic structure is a sequence of emitting signals from bottom to top:
enemy -> SIG: died -> level -> SIG: scored -> main -> $HUD.set_score

IDEA: This way later the level can influence the amount of points e.g. depending on the type of enemy that died and the difficulty of the level.